added totp(mfa), note, trading view and mails

+++ src/main/java/com/fintech/crypto/service/utility/PasswordResetSvc.java
@@ -76,17 +76,21 @@ public class PasswordResetSvc implements PasswordResetCt {
     @Override
     public Boolean resetPassword(PasswordResetReq request) {
 
-        Optional<PasswordResetSession> checkedToken = passwordResetSessionDao.findByToken(request.getVerificationToken());
+        if (!request.getPassword().equals(request.getConfirmPassword())){
+            throw new UnsatisfiedLinkError("Password does not match");
+        }
+
+        Optional<PasswordResetSession> checkedToken = passwordResetSessionDao.findByToken(request.getResetToken());
         if(checkedToken.isPresent()){
 
             //Check if token is expired
             if( checkedToken.get().getCreatedAt().plusSeconds(Long.parseLong(prop.PASSWORD_VERIFICATION_THRESHOLD)).isBefore(LocalDateTime.now())  ) {
                 passwordResetSessionDao.delete(checkedToken.get());
                 passwordResetSessionDao.flush();
-                throw new RuntimeException(String.format("Verification token %s has expired, exceeds %d min", request.getVerificationToken(), Long.parseLong(prop.PASSWORD_VERIFICATION_THRESHOLD) / 60));
+                throw new RuntimeException(String.format("Reset link %s has expired, exceeds %d min", request.getResetToken(), Long.parseLong(prop.PASSWORD_VERIFICATION_THRESHOLD) / 60));
             }
         }else{
-            throw new IllegalStateException(String.format("Verification token %s not found in data store", request.getVerificationToken()));
+            throw new IllegalStateException(String.format("Reset link %s not found in data store", request.getResetToken()));
         }
         //Get email attached
         String email = checkedToken.get().getEmail();
@@ -94,7 +98,7 @@ public class PasswordResetSvc implements PasswordResetCt {
                 () -> {
                     throw new UsernameNotFoundException(
                             String.format(
-                                    "Verification token not found for %s",
+                                    "Reset link not found for %s",
                                     email
                             )
                     );
